fix: Support jumping between node.js versions
Use `NODE_PATH` and use bare filename for `--require` if it is needed in
node.js < 12 or if node.js 12+ would otherwise require quoting the
require path.  This should avoid a crash when `node-preload` is setup in
node.js 12 but then a test is run under node.js < 12.  It should also
ensure compatibility when setup is performed in node.js < 12 and tests
are later run under node.js 12+.

Ref istanbuljs/nyc#1246
