--- generate-require-modern.js
@@ -1,12 +1,0 @@
-'use strict';
-
-function generateRequire(filename) {
-	return `--require "${filename.replace(/\\/g, '\\\\').replace(/"/g, '\\"')}"`;
-}
-
-module.exports = {
-	generateRequire,
-	// The export explicitly exists but is undefined as it will never be called
-	processNodePath: undefined,
-	needsPathEnv: () => false
-};
