+++ generate-require.js
@@ -1,6 +1,31 @@
 'use strict';
 
-/* istanbul ignore next: version specific branching */
-const requireType = Number(process.versions.node.split('.')[0]) < 12 ? 'legacy' : 'modern';
+const path = require('path');
 
-module.exports = require(`./generate-require-${requireType}.js`);
+const needsPathRegExp = /[\\ "]/;
+
+const needsPathEnv = dir => needsPathRegExp.test(dir);
+
+function generateRequire(filename) {
+	if (needsPathEnv(filename)) {
+		return `--require ${path.basename(filename)}`;
+	}
+
+	return `--require ${filename}`;
+}
+
+function processNodePath(value) {
+	const dir = path.dirname(require.resolve('./preload-path/node-preload.js'));
+	const existing = value === '' ? [] : value.split(path.delimiter);
+	if (existing.includes(dir)) {
+		return value;
+	}
+
+	return existing.concat(dir).join(path.delimiter);
+}
+
+module.exports = {
+	generateRequire,
+	processNodePath,
+	needsPathEnv
+};
